Added default dictionary & changed file structure
In case things go wrong reading dictionary.txt, a hard coded default dictionary with the category of countries is loaded (it is the same as the one found in /src/dictionary.txt).  Also, I imported the whole project into IntelliJ so that the code would be easier to work with and compile (a copy of acm.jar is also included).